refactor(channels): split controllers/ops.rs into ops/ directory
Splits the 1333-line ops.rs into six focused sub-modules (each ≤500 lines):
types, connect, yuanbao, telegram, discord, messaging. External API preserved
via pub use re-exports in mod.rs.
